be/requirements.txt
beanie 
datasets 
fastapi 
fastapi_mail
langchain
langchain_community 
langchain_text_splitters 
motor 
openai 
pandas 
pydantic 
groq
pydantic_settings 
python-dotenv 
redis[hiredis]
tqdm 
uvicorn 
PyJWT
Jinja2
python-multipart
weaviate-client
websockets
minio
langchain-pymupdf4llm
tiktoken
prometheus-fastapi-instrumentator
langchain_openai
passlib[bcrypt]